% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvLRHE.R
\name{cv_mvLRHE_L2}
\alias{cv_mvLRHE_L2}
\title{cv_mvLRHE_rank}
\usage{
cv_mvLRHE_L2(
  Y,
  D_list,
  K,
  n_lambda = 5,
  lambda_max = 1e-04,
  lambda_min = 1e-12,
  tolerance = 1e-07,
  max_iter = 10000,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
cv_mvLRHE_rank
}
