% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{cv_mvREHE}
\alias{cv_mvREHE}
\title{mvREHE}
\usage{
cv_mvREHE(
  Y,
  D_list,
  K,
  n_lambda = 10,
  lambda_max = 1,
  lambda_min = 1e-08,
  tolerance = 1e-09,
  max_iter = 10000,
  L_init_list = NULL,
  algorithm = "L-BFGS-B"
)
}
\arguments{
\item{algorithm}{}
}
\description{
mvREHE
}
