% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{cv_mvREHE_L2}
\alias{cv_mvREHE_L2}
\title{cv_mvREHE_L2}
\usage{
cv_mvREHE_L2(
  Y,
  D_list,
  K = 5,
  folds = NULL,
  grid = FALSE,
  n_lambda = 10,
  lambda_max = 0.001,
  lambda_min = 1e-06,
  tolerance = NULL,
  max_iter = 100,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
cv_mvREHE_L2
}
