% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{cv_mvREHE_rank}
\alias{cv_mvREHE_rank}
\title{cv_mvREHE_rank}
\usage{
cv_mvREHE_rank(
  Y,
  D_list,
  K,
  n_rank = 5,
  rank_max = ncol(Y),
  rank_min = 1,
  tolerance = 1e-07,
  max_iter = 10000,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
cv_mvREHE_rank
}
