% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvHE_diag.R
\name{mvHE_diag}
\alias{mvHE_diag}
\title{mvHE_diag}
\usage{
mvHE_diag(Y, D_list, W_row_pairs = NULL, w_columns = NULL, truncate = TRUE)
}
\description{
Like mvHE but uses only same-observation pairs (diagonal of D_k) in the
moment equations: G_k = Y' diag(D_k) Y, Q[k,z] = sum(diag(D_k)*diag(D_z)).
Equivalent to equation 13 of the mvREHE paper with i=l only.
}
