\name{mvREHE-package}
\alias{mvREHE-package}
\alias{mvREHE}
\docType{package}
\title{
\packageTitle{mvREHE}
}
\description{
\packageDescription{mvREHE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mvREHE}
\packageIndices{mvREHE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mvREHE}

Maintainer: \packageMaintainer{mvREHE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
