% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE_cvDR}
\alias{mvREHE_cvDR}
\title{mvREHE_cvDR}
\usage{
mvREHE_cvDR(
  Y,
  D_list,
  r_seq = NULL,
  V_function = svd_irlba,
  K = 5,
  folds = NULL,
  compute_full_Sigma = FALSE,
  tolerance = 0.001,
  max_iter = 100
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
mvREHE_cvDR
}
