% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE_cvL2}
\alias{mvREHE_cvL2}
\title{mvREHE_cvL2}
\usage{
mvREHE_cvL2(
  Y,
  D_list,
  K = 5,
  folds = NULL,
  grid = TRUE,
  n_lambda = 5,
  lambda_max = 0.01,
  lambda_min = 1e-06,
  tolerance = 0.001,
  max_iter = 100,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
mvREHE_cvL2
}
