% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvHE_diag.R
\name{mvREHE_diag}
\alias{mvREHE_diag}
\title{mvREHE_diag}
\usage{
mvREHE_diag(
  Y,
  D_list,
  W_row_pairs = NULL,
  w_columns = NULL,
  tolerance = 1e-06,
  max_iter = 1000,
  return_full = TRUE,
  Sigma_init_list = NULL,
  track_loss = FALSE,
  truncate = TRUE
)
}
\description{
Like mvREHE but uses only same-observation pairs (diagonal of D_k) in the
moment equations: G_k = Y' diag(D_k) Y, Q[k,z] = sum(diag(D_k)*diag(D_z)).
}
