% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE_DR}
\alias{mvREHE_DR}
\title{mvREHE_DR}
\usage{
mvREHE_DR(
  Y,
  D_list,
  V = NULL,
  tolerance = NULL,
  max_iter = 100,
  Sigma_init_list = NULL,
  Q = NULL,
  row_indices = NULL,
  col_indices = NULL
)
}
\arguments{
\item{col_indices}{}
}
\description{
mvREHE_DR
}
