% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvLRHE.R
\name{oracle_mvLRHE}
\alias{oracle_mvLRHE}
\title{oracle_mvLRHE}
\usage{
oracle_mvLRHE(
  Y,
  D_list,
  Sigma_list,
  n_rank = 10,
  rank_max = ncol(Y),
  rank_min = 1,
  tolerance = 1e-09,
  max_iter = 10000,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
oracle_mvLRHE
}
