% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Perform massively univariate analysis}
\usage{
univariate(Y, D_list, estimator)
}
\arguments{
\item{estimator}{}
}
\description{
Perform massively univariate analysis
}
