% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{cv_mvREHE_L2}
\alias{cv_mvREHE_L2}
\title{cv_mvREHE_rank}
\usage{
cv_mvREHE_L2(
  Y,
  D_list,
  K,
  n_lambda = 3,
  lambda_max = 1e-06,
  lambda_min = 1e-10,
  tolerance = 1e-07,
  max_iter = 10000,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
cv_mvREHE_rank
}
