% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE}
\alias{mvREHE}
\title{mvREHE}
\usage{
mvREHE(
  Y,
  D_list,
  tolerance = 1e-06,
  max_iter = 1000,
  return_full = TRUE,
  Sigma_init_list = NULL
)
}
\arguments{
\item{col_indices}{}
}
\description{
mvREHE
}
