% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE}
\alias{mvREHE}
\title{mvREHE}
\usage{
mvREHE(
  Y,
  D_list,
  r = NULL,
  lambda = NULL,
  tolerance = 1e-07,
  max_iter = 10000,
  Sigma_init_list = NULL
)
}
\arguments{
\item{Sigma_init_list}{}
}
\description{
mvREHE
}
