% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE.R
\name{mvREHE}
\alias{mvREHE}
\title{mvREHE}
\usage{
mvREHE(
  Y,
  D_list,
  tolerance = 1e-09,
  max_iter = 10000,
  L_init_list = NULL,
  algorithm = "GD"
)
}
\arguments{
\item{algorithm}{}
}
\description{
mvREHE
}
