% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE_SNP.R
\name{mvREHE_SNP}
\alias{mvREHE_SNP}
\title{mvREHE_SNP}
\usage{
mvREHE_SNP(
  Y,
  D_list,
  GWAS_N,
  init_h2 = 0.1,
  init_rho_g = 0.3,
  init_rho_e = 0.05,
  tolerance = 1e-06,
  max_iter = 1000,
  return_full = TRUE,
  Sigma_init_list = NULL
)
}
\arguments{
\item{col_indices}{}
}
\description{
mvREHE_SNP
}
