% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvREHE_SNP.R
\name{mvREHE_SNP}
\alias{mvREHE_SNP}
\title{mvREHE_SNP}
\usage{
mvREHE_SNP(
  Y,
  D_list,
  GWAS_N,
  M = NULL,
  init = c("mvHE", "mvREHE"),
  truncate = TRUE,
  cross_snp = FALSE
)
}
\description{
mvREHE_SNP
}
